requirements.txt
torch
torchvision
huggingface_hub
datasets
earthengine-api
rasterio
pandas
numpy
Pillow
PyYAML
tqdm